chore(deps): cleanup
