--- function/test/hello.test.ts
@@ -1,5 +1,0 @@
-import { Hello } from '../src';
-
-test('hello', () => {
-  expect(new Hello().sayHello()).toBe('hello, world!');
-});
